api/requirements.txt
fastapi
uvicorn
yt-dlp
moviepy
openai
python-dotenv